big change

+++ include/merlin/aloctr.h
@@ -37,7 +37,7 @@ struct mrln_aloctr_layout_t {
 __attribute__((__always_inline__, __nonnull__(1, 2, 3),
                __warn_unused_result__)) static inline int
 mrln_alloc(mrln_aloctr_t *a, void **chunk, intptr_t *chunk_size,
-           const intptr_t align, const intptr_t size) {
+           const int align, const intptr_t size) {
   return ((mrln_aloctr_layout_t *)a)->alloc(a, chunk, chunk_size, align, size);
 }
